+++ ReleaseNotes/release_notes.md
@@ -7,7 +7,8 @@
   1. reduce code duplication between  GoF-1D and Two-Sample Tests
   1. Add Anderson-Darling and Cramer-von-Mises Two Sample Tests
   1. Add progress bar to methdos dealing with unique/duplicated entries in dataset
-    
+  1. Add Hilbert transform
+        
 ## Backward incompatible
 
 ## Bug fixes